+++ NEWS.md
@@ -1,3 +1,17 @@
+# parameters 0.18.1
+
+## Changes to functions
+
+* `model_parameters()` now also includes standard errors and confidence
+  intervals for slope-slope-correlations of random effects variances.
+
+* `ci()` for *glmmTMB* models with `method = "profile"` is now more robust.
+
+## Bug fixes
+
+* Fixed issue with *glmmTMB* models when calculating confidence 
+  intervals for random effects failed due to singular fits.
+
 # parameters 0.18.0
 
 ## Breaking Changes